Increase maximum allowable API ID in Settings
The NiceHashBot Settings is a Windows Form application that uses System.Windows.Forms.NumericUpDown form, which requires that some maximum value is hard coded. The current limit is 1000000 (one million).
As of 2018, new NiceHash users are assigned API ID values in excess of 1000000 (there are probably now more than one million nicehash users). As a result, new user accounts cannot be used with NiceHashBot because their API ID cannot be entered.

Because the existing maximum value is too low, I propose making it much larger.  MaxValue for the decimal class can accept a value that is very high, so in the code above, I have switched the max value to one trillion instead of one million.